pyproject.toml
[project]
name = "mission-control-demo"
version = "0.1.0"
description = "Mission Control — a voice assistant and MCP operations demo"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp",
    "httpx",
    "livekit-agents[deepgram,groq,openai,sarvam,silero]>=1.5.1",
    "livekit-plugins-google>=1.5.1",
    "python-dotenv",
]

[project.scripts]
mission_control = "mission_control_server:main"
mission_control_voice = "mission_control_agent:dev"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["mission_control"]
